// .kilo/worktrees/ordinary-poultry/src/services/socket/socket.types.ts
export interface ChatRequestPayload {
  name: string;
  user_id: string;
  astro_id: string;
  room_id: string;
  maximum_time: number;
  phoneNumber: string;
  userName: string;
  gender: string;
  timeOfBirth: string;
  location: string;
  dateOfBirth?: string;
  birthTime?: string;
  is_promotional?: boolean;
  user_image?: string;
  occupation?: string;
}